// lib/api/generated/model/milestoneCreate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Syncnesto API
 * OpenAPI spec version: 0.1.0
 */

/**
 * マイルストーン作成リクエストで受け取るschema。
 */
export interface MilestoneCreate {
  title: string;
  description?: string | null;
  target_date: string;
  status?: string;
}